+++ src/main/java/com/foxminded/university/management/schedule/service/impl/StudentServiceImpl.java
@@ -1,10 +1,10 @@
 package com.foxminded.university.management.schedule.service.impl;
 
-import com.foxminded.university.management.schedule.exceptions.ServiceException;
 import com.foxminded.university.management.schedule.models.Student;
 import com.foxminded.university.management.schedule.repository.GroupRepository;
 import com.foxminded.university.management.schedule.repository.StudentRepository;
 import com.foxminded.university.management.schedule.service.StudentService;
+import com.foxminded.university.management.schedule.service.exceptions.EntityNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -33,7 +33,7 @@ public class StudentServiceImpl implements StudentService {
         if (isGroupPresent || student.getGroup() == null) {
             return studentRepository.save(student);
         }
-        throw new ServiceException("Student's group with id: " + student.getGroup() + " is not exist");
+        throw new EntityNotFoundException("Student's group with id: " + student.getGroup() + " is not exist");
     }
 
     @Override
@@ -43,7 +43,7 @@ public class StudentServiceImpl implements StudentService {
         if (isStudentPresent) {
             return studentRepository.findById(id).get();
         }
-        throw new ServiceException("Student with id: " + id + " is not found");
+        throw new EntityNotFoundException("Student with id: " + id + " is not found");
     }
 
     @Override
@@ -53,7 +53,13 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void deleteStudentById(Long id) {
-        studentRepository.deleteById(getStudentById(id).getId());
+        boolean isStudentPresent = studentRepository.findById(id).isPresent();
+        LOGGER.debug("Student is present: {}", isStudentPresent);
+        if (isStudentPresent) {
+            studentRepository.deleteById(getStudentById(id).getId());
+        } else {
+            throw new EntityNotFoundException("Student with id: " + id + " is not found");
+        }
     }
 
     @Override
